Added DirectionsHelper Added Launch Screen Added App Icon assets
The directions helper allows you to hand it a start, end, n intermediate locations, and a completion
handler. It uses googles directions service to get the optimized route given the intermediate
locations (Traveling Salesman problem woot woot!). In the completion handler if the line is present
just set its map to the map you want it displayed on and it shows up.

There are also a few snippets left in as examples for things like the places service which we may
find useful in the future.

Launch screen thoughts? the symbol is pretty cool!

![Video Walkthrough](RouteOptimizerWIPDemo.gif)